Switch to a fork of trust-dns-resolver to fix an integer overflow
In Version 0.9.0 (660ca00c9b9ff0cbd675) of trust-dns-resolver there is
an integer overflow on 32-bit processor architectures.
This bug was fixed in trust-dns in commit 11de5183c330b2. We use the fork
of trust-dns 0.9.0 which cherry-picks this fix until upstream
linkerd2-proxy switches to a newer trust-dns-resolver version that
includes this fix.